+++ backend/ms-seguridad/src/auth/strategy.ts
@@ -6,7 +6,7 @@ import { inject, service } from '@loopback/core';
 import {HttpErrors, Request} from '@loopback/rest';
 import {UserProfile} from '@loopback/security';
 import parseBearerToken from 'parse-bearer-token';
-import {SeguridadUsuarioService} from '../services';
+import {AuthService, SeguridadUsuarioService} from '../services';
 import {RolMenuRepository} from '../repositories';
 import { repository } from '@loopback/repository';
 
@@ -21,6 +21,8 @@ export class AuthStrategy implements AuthenticationStrategy {
     private metadata: AuthenticationMetadata[],
     @repository(RolMenuRepository)
     private rolMenuRepository : RolMenuRepository,
+    @service(AuthService)
+    private servicioAuth: AuthService,
     ) {}
   
 
@@ -36,44 +38,15 @@ export class AuthStrategy implements AuthenticationStrategy {
       let idMenu:string = this.metadata[0].options![0];
       let accion:string = this.metadata[0].options![1];
       console.log(this.metadata);
-      let permiso = await this.rolMenuRepository.findOne({
-        where:{
-          rolId:idRol,
-          menuId:idMenu,
-        }
-      });
-      let continuar : boolean = false;
-      if(permiso){
-        switch(accion){
-           case "guardar":
-            continuar = permiso.guardar;
-            break;
-          case "editar":
-            continuar = permiso.editar;
-            break;
-          case "listar":
-            continuar = permiso.listar;
-            break;
-          case "eliminar":
-            continuar = permiso.eliminar;
-            break;
-          case "descargar":
-            continuar = permiso.descargar;
-              break;
-          default:
-            throw new HttpErrors[401]("no es posible ejecutar la acción porque no existe")
-        };
-        if(continuar){
-           let perfil: UserProfile = Object.assign({
-            permitido : "OK"
-           });
-           return perfil;
-        }else{
-          return undefined;
-        }
-        }
-      else{
-        throw new HttpErrors[401]("no tiene permisos")};
+      /**Se manda la funcion anterior al auth service y ahora lo consumimos de esta forma aqui, importamos 
+       * el auth service y lo inyectamos en el constructor y llamamos la funcion desde el servicio.
+      */
+      try{
+        let res = await this.servicioAuth.VerificarPermisoDeUsuarioPorRol(idRol, idMenu, accion);
+        return res;
+      }catch(e){
+        throw e;
+      }
      }
      throw new HttpErrors[401]("no se puede ejecutar la acción por falta de un token.");
     }
