+++ backend/ms-seguridad/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './seguridad-usuario.service';
+export * from './auth.service';
